+++ src/run/driver/mod.rs
@@ -3,6 +3,8 @@
 //!
 //
 
+mod _helper; // _run_driver_step_run_frame_body!
+
 mod driver; // RunDriver
 mod error; // RunDriverError, RunDriverFrameError
 
@@ -14,6 +16,6 @@ crate::structural_mods! { // _mods, _crate_internals
         };
     }
     _crate_internals {
-        pub(crate) use super::driver::_run_driver_step_frame_body;
+        pub(crate) use super::_helper::*;
     }
 }